+++ CHANGELOG.md
@@ -5,5 +5,6 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Protocol for unmanaged resources, defining a method to free the resource
 - Initialize GLFW
+- Create windows
 
 [Unreleased]: https://github.com/IGJoshua/s-expresso/compare/master..develop